+++ crates/orchestrator/src/database/mongodb/mod.rs
@@ -39,7 +39,11 @@ impl MongoDb {
         // Get a handle to the cluster
         let client = Client::with_options(client_options).expect("Failed to create MongoDB client");
         // Ping the server to see if you can connect to the cluster
-        client.database("admin").run_command(doc! {"ping": 1}, None).await.expect("Failed to ping MongoDB deployment");
+        client
+            .database("orchestrator")
+            .run_command(doc! {"ping": 1}, None)
+            .await
+            .expect("Failed to ping MongoDB deployment");
         tracing::debug!("Pinged your deployment. You successfully connected to MongoDB!");
 
         Self { client, database_name: mongo_db_settings.database_name }