+++ crates/orchestrator/src/jobs/state_update_job/mod.rs
@@ -130,16 +130,19 @@ impl Job for StateUpdateJob {
         let mut sent_tx_hashes: Vec<String> = Vec::with_capacity(block_numbers.len());
         for block_no in block_numbers.iter() {
             tracing::debug!(job_id = %job.internal_id, block_no = %block_no, "Processing block");
+
             let snos = self.fetch_snos_for_block(*block_no, config.clone()).await;
-            let tx_hash = self.update_state_for_block(config.clone(), *block_no, snos, nonce).await.map_err(|e| {
-                tracing::error!(job_id = %job.internal_id, block_no = %block_no, error = %e, "Error updating state for block");
-                job.metadata.insert(JOB_METADATA_STATE_UPDATE_LAST_FAILED_BLOCK_NO.into(), block_no.to_string());
-                self.insert_attempts_into_metadata(job, &attempt_no, &sent_tx_hashes);
-                StateUpdateError::Other(OtherError(eyre!(
-                    "Block #{block_no} - Error occurred during the state update: {e}"
-                )))
-            })?;
-            sent_tx_hashes.push(tx_hash);
+            let txn_hash = self
+                .update_state_for_block(config.clone(), *block_no, snos, nonce)
+                .await
+                .map_err(|e| {
+                    tracing::error!(job_id = %job.internal_id, block_no = %block_no, error = %e, "Error updating state for block");
+                    job.metadata.insert(JOB_METADATA_STATE_UPDATE_LAST_FAILED_BLOCK_NO.into(), block_no.to_string());
+                    self.insert_attempts_into_metadata(job, &attempt_no, &sent_tx_hashes);
+                    OtherError(eyre!("Block #{block_no} - Error occurred during the state update: {e}"));
+                })
+                .unwrap();
+            sent_tx_hashes.push(txn_hash);
             nonce += 1;
         }
 
@@ -318,6 +321,7 @@ impl StateUpdateJob {
                 .map_err(|e| JobError::Other(OtherError(e)))?;
 
             let program_output = self.fetch_program_output_for_block(block_no, config.clone()).await;
+
             // TODO :
             // Fetching nonce before the transaction is run
             // Sending update_state transaction from the settlement client