+++ crates/orchestrator/src/workers/proof_registration.rs
@@ -1,4 +1,3 @@
-use std::error::Error;
 use std::sync::Arc;
 
 use async_trait::async_trait;
@@ -13,7 +12,7 @@ impl Worker for ProofRegistrationWorker {
     /// 1. Fetch all blocks with a successful proving job run
     /// 2. Group blocks that have the same proof
     /// 3. For each group, create a proof registration job with from and to block in metadata
-    async fn run_worker(&self, _config: Arc<Config>) -> Result<(), Box<dyn Error>> {
+    async fn run_worker(&self, _config: Arc<Config>) -> color_eyre::Result<()> {
         todo!()
     }
 }