+++ crates/orchestrator/src/workers/proving.rs
@@ -1,4 +1,3 @@
-use std::error::Error;
 use std::sync::Arc;
 
 use async_trait::async_trait;
@@ -14,7 +13,7 @@ pub struct ProvingWorker;
 impl Worker for ProvingWorker {
     /// 1. Fetch all successful SNOS job runs that don't have a proving job
     /// 2. Create a proving job for each SNOS job run
-    async fn run_worker(&self, config: Arc<Config>) -> Result<(), Box<dyn Error>> {
+    async fn run_worker(&self, config: Arc<Config>) -> color_eyre::Result<()> {
         tracing::trace!(log_type = "starting", category = "ProvingWorker", "ProvingWorker started.");
 
         let successful_snos_jobs = config
@@ -26,12 +25,7 @@ impl Worker for ProvingWorker {
 
         for job in successful_snos_jobs {
             tracing::debug!(job_id = %job.internal_id, "Creating proof creation job for SNOS job");
-            match create_job(JobType::ProofCreation, job.internal_id.to_string(), job.metadata.clone(), config.clone())
-                .await
-            {
-                Ok(_) => tracing::info!(block_no = %job.internal_id, "Successfully created proof creation job"),
-                Err(e) => tracing::error!(job_id = %job.internal_id, error = %e, "Failed to create proof creation job"),
-            }
+            create_job(JobType::ProofCreation, job.internal_id.to_string(), job.metadata, config.clone()).await?;
         }
 
         tracing::trace!(log_type = "completed", category = "ProvingWorker", "ProvingWorker completed.");